Fixed bug of disappearing card and fixed question loading
Bug 1: After swiping bottom card, it would disappear and not reload. This
was caused by a misconstructed if statement.

Bug 2: After loading the first round of questions, we would stop pulling
more questions from our list. This was also caused by a careless IF
statment.

Bug 3: Tweaked slide threshold so that answers are selected earlier in
movement

Todo: QUestions are being loaded a few times over - need to fix that.